backend/requirements.txt
flask>=1.1.1
flask_cors
google-cloud-storage
google-cloud-bigquery
gsutil
firebase-admin